Add gradient computation feature (#2)
* Implemented `nabla` and `nabla_t`

In all honesty, the sole purpose of `nabla_t` is to support similar
signatures to what is required by nyx::propagators::Propagator.derive .

* Moved the differentiating functions to their own submodule

* WIP

* Revert changes to Cargo.toml -- bumped version to 0.2.0

Version bump due to breaking change of DualNumber to Dual.

* First throw at implementing some linear algebra functions

I can't just derive them for MatrixMN because that comes from another
crate. I also can't make Dual a vector space in itself. I am not 100%
sure this is mathematically correct. I tried, but that led to a
ridiculous number of changes, and it still didn't work.

* Add gradient computation as a feature